+++ R/prblty.R
@@ -101,9 +101,17 @@ prblty <- function(link, p) {
     # Many values will have "X" for less than 1% chance. Make 0
     prblty[prblty == "X"] <- 0
 
-    prblty <- dplyr::mutate_at(.tbl = prblty,
-                               .cols = c(2:6),
-                               .funs = "as.numeric")
+    # dplyr 0.6.0 renames .cols parameter to .vars. For the time being,
+    # accomodate usage of both 0.5.0 and >= 0.6.0.
+    if (packageVersion("dplyr") > "0.5.0") {
+        prblty <- dplyr::mutate_at(.tbl = prblty,
+                                   .vars = c(2:6),
+                                   .funs = "as.numeric")
+    } else {
+        prblty <- dplyr::mutate_at(.tbl = prblty,
+                                   .cols = c(2:6),
+                                   .funs = "as.numeric")
+    }
 
     prblty <- prblty %>%
         dplyr::mutate("Status" = status,